% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_ata.R
\name{components.ATA}
\alias{components.ATA}
\title{Extract estimated states from an ATA model.}
\usage{
\method{components}{ATA}(object, ...)
}
\arguments{
\item{object}{An estimated model.}

\item{...}{Unused.}
}
\value{
A [fabletools::dable()] containing estimated states.
}
\description{
Extract estimated states from an ATA model.
}
\examples{
library(fable.ata)
as_tsibble(USAccDeaths) \%>\%
  model(ata = AutoATA(value ~ trend("A") + season("A"))) \%>\% components()

}
