% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_ata.R
\name{components.fable_ata}
\alias{components.fable_ata}
\title{Extract estimated states from an ATA model.}
\usage{
\method{components}{fable_ata}(object, ...)
}
\arguments{
\item{object}{An estimated model.}

\item{...}{Unused.}
}
\value{
A [fabletools::dable()] containing estimated states.
}
\description{
Extract estimated states from an ATA model.
}
