% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.ata-package.R, R/fable_ata.R
\docType{package}
\name{fable.ata-package}
\alias{fable.ata}
\alias{fable.ata-package}
\title{fable.ata: Fable Modelling Wrappers for ATAforecasting Package}
\description{
Allows Ata models from the 'ATAforecasting' package to be used in a tidy workflow with the modelling interface of 'fabletools'. This extends 'ATAforecasting' to provide enhanced model specification and management, performance evaluation methods, and model combination tools.

Allows Ata models from the 'ATAforecasting' package to be used in a tidy workflow with the modelling interface of 'fabletools'. This extends 'ATAforecasting' to provide enhanced model specification and management, performance evaluation methods, and model combination tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alsabtay/fable.ata https://atamethod.wordpress.com/}
  \item Report bugs at \url{https://github.com/alsabtay/fable.ata/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/alsabtay/fable.ata https://atamethod.wordpress.com/}
  \item Report bugs at \url{https://github.com/alsabtay/fable.ata/issues}
}

}
\author{
\strong{Maintainer}: Ali Sabri Taylan \email{alisabritaylan@gmail.com} (\href{https://orcid.org/0000-0001-9514-934X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Hanife Taylan Selamlar (\href{https://orcid.org/0000-0002-4091-884X}{ORCID}) [copyright holder]
  \item Guckan Yapar (\href{https://orcid.org/0000-0002-0971-6676}{ORCID}) [thesis advisor, copyright holder]
}


\strong{Maintainer}: Ali Sabri Taylan \email{alisabritaylan@gmail.com} (\href{https://orcid.org/0000-0001-9514-934X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Hanife Taylan Selamlar (\href{https://orcid.org/0000-0002-4091-884X}{ORCID}) [copyright holder]
  \item Guckan Yapar (\href{https://orcid.org/0000-0002-0971-6676}{ORCID}) [thesis advisor, copyright holder]
}

}
\keyword{internal}
\keyword{package}
