% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_ata.R
\name{fitted.ATA}
\alias{fitted.ATA}
\title{Extract fitted values}
\usage{
\method{fitted}{ATA}(object, ...)
}
\arguments{
\item{object}{An estimated model.}

\item{...}{Unused.}
}
\value{
A vector of fitted values.
}
\description{
Extracts the fitted values from an estimated ATA model.
}
