% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_ata.R
\name{format.ATA}
\alias{format.ATA}
\title{Format of ATA model}
\usage{
\method{format}{ATA}(x, ...)
}
\arguments{
\item{x}{An estimated model.}

\item{...}{Unused.}
}
\value{
The forecasting model's name.
}
\description{
Format of ATA model
}
