% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_ata.R
\name{format.fable_ata}
\alias{format.fable_ata}
\title{Format of ATA model}
\usage{
\method{format}{fable_ata}(x, ...)
}
\arguments{
\item{x}{An estimated model.}

\item{...}{Unused.}
}
\description{
Format of ATA model
}
