% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_ata.R
\name{glance.fable_ata}
\alias{glance.fable_ata}
\title{Glance an ATA model}
\usage{
\method{glance}{fable_ata}(x, ...)
}
\arguments{
\item{x}{An estimated model.}

\item{...}{Unused.}
}
\value{
A one row tibble summarising the model's fit.
}
\description{
Glance an ATA model
}
