% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_ata.R
\name{model_sum.ATA}
\alias{model_sum.ATA}
\title{Summary of ATA model}
\usage{
\method{model_sum}{ATA}(x, ...)
}
\arguments{
\item{x}{An estimated model.}

\item{...}{Unused.}
}
\value{
The model's summary specs in a `tibble`.
}
\description{
Summary of ATA model
}
