% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_ata.R
\name{model_sum.fable_ata}
\alias{model_sum.fable_ata}
\title{Summary of ATA model}
\usage{
\method{model_sum}{fable_ata}(x, ...)
}
\arguments{
\item{x}{An estimated model.}

\item{...}{Unused.}
}
\value{
The model's summary specs.
}
\description{
Summary of ATA model
}
