% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_ata.R
\name{report.ATA}
\alias{report.ATA}
\title{Specialized Screen Print Function of ATA model}
\usage{
\method{report}{ATA}(object, ...)
}
\arguments{
\item{object}{An estimated model.}

\item{...}{Unused.}
}
\value{
a summary for the results of the ATAforecasting
}
\description{
Specialized Screen Print Function of ATA model
}
\examples{
 library(fable.ata)
 as_tsibble(USAccDeaths) \%>\% model(ata = AutoATA(value ~ trend("A") + season("M"))) \%>\% report()

}
