% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_ata.R
\name{report.fable_ata}
\alias{report.fable_ata}
\title{Specialized Screen Print Function of ATA model}
\usage{
\method{report}{fable_ata}(object, ...)
}
\arguments{
\item{object}{An estimated model.}

\item{...}{Unused.}
}
\description{
Specialized Screen Print Function of ATA model
}
