% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_ata.R
\name{residuals.fable_ata}
\alias{residuals.fable_ata}
\title{Extract model residuals}
\usage{
\method{residuals}{fable_ata}(object, ...)
}
\arguments{
\item{object}{An estimated model.}

\item{...}{Unused.}
}
\value{
A vector of residuals.
}
\description{
Extracts the residuals from an estimated ATA model.
}
