% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_ata.R
\name{tidy.ATA}
\alias{tidy.ATA}
\title{Tidy a ATA model}
\usage{
\method{tidy}{ATA}(x, ...)
}
\arguments{
\item{x}{An estimated model.}

\item{...}{Unused.}
}
\value{
The model's coefficients in a `tibble`.
}
\description{
Tidy a ATA model
}
\examples{
as_tsibble(USAccDeaths) \%>\%
  model(ata = AutoATA(value ~ trend("A") + season("M"))) \%>\% tidy()

}
