% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_ata.R
\name{tidy.fable_ata}
\alias{tidy.fable_ata}
\title{Tidy a ATA model}
\usage{
\method{tidy}{fable_ata}(x, ...)
}
\arguments{
\item{x}{An estimated model.}

\item{...}{Unused.}
}
\value{
The model's coefficients in a `tibble`.
}
\description{
Tidy a ATA model
}
